libstdc++: Update personality function for purecap
This updates the C++ personality function following the changes made to
the libgcc C personality function for pure-capability Morello. There are
two main changes here. The first is using
__builtin_code_address_from_pointer to drop the LSB from Morello code
pointers when searching through call-site tables (without this we would
never find the right landing pad when unwinding).

The second change is to reflect the change in the exception table format
for pure-capability Morello where the landing pad is a capability
indirected by an offset in the call-site table.

We also improve on the robustness of the existing code in the C
personality function by asserting that we only see the implemented
capability landing pad marker value (0xd) and also handling the case
where this is no associated landing pad with a given call-site table
entry.

This fixes many execution tests involving exception handling in the g++
testsuite.

libgcc/ChangeLog:

	* unwind-c.c (PERSONALITY_FUNCTION): Fix handling of case where
	call-site table entry doesn't have an associated landing pad.

libstdc++-v3/ChangeLog:

	* libsupc++/eh_personality.cc (PERSONALITY_FUNCTION): Adapt for
	pure-capability Morello.