middle-end: Fix ICE in tree-ssa-forwprop
This fixes an ICE in tree-ssa-forwprop.c:forward_propagate_addr_expr_1
which was due to a case of TYPE_PRECISION not being updated to one of
TYPE_[NON]CAP_PRECISION. Here we assert that for capability types, the
two types must have the same precision. For non-capability types, we
preserve the behaviour of the existing code (recursing on widening or
equal-precision conversions), but use TYPE_NONCAP_PRECISION to avoid
tripping the check when we handle capability types.

gcc/ChangeLog:

	* tree-ssa-forwprop.c (forward_propagate_addr_expr_1): Update
	TYPE_PRECISION check to handle capability types.

gcc/testsuite/ChangeLog:

	* gcc.target/aarch64/morello/forwprop-ice.c: New test.