cp: Fix pointer comparison bug with capabilities
This fixes a wrong code bug when comparing capability pointers. In
cp_build_binary_op we would of course do pointer comparison in the
non-capability type to get the correct CHERI C/C++ semantics, so we
would just directly convert both operands to the non-capability type.
This means we miss the conversion via the common pointer type, which is
problematic when this conversion is not a no-op (e.g. converting from a
pointer to a derived class to a pointer to a base class with multiple
inheritance where a "this adjustment" is necessary).

This patch fixes the bug by first converting to the common pointer type
before converting to the final non-capability result type.