fold-const.c: Handle capability pointer comparisons
This handles folding comparisons of capability pointers in
fold-const.c:fold_comparison. This is needed for some C++ constexpr
tests to pass. E.g. g++.dg/cpp0x/constexpr-67376.C and
g++.dg/cpp0x/constexpr-array-ptr9.C.

When the FE forms comparisons on capabilities, it does so by first
casting to the non-capability type of the capabilities and performing
the comparison on the address values, so the comparisons are of the
form:

(CV) c1 <cmp> (CV) c2

where CV is the non-capability type of the capabilities. As it stands,
fold_comparison is unable to fold such comparisons. With this patch, we
pull out all the existing logic for folding pointer-typed expressions
into a new function, fold_address_value_comparison, which can be used to
fold comparisons of any two pointer-typed expressions (capabilities or
otherwise), preserving the semantics that we compare the address values
of the capabilities (not the capabilities themselves).

We then adjust fold_comparison to recognize capability comparisons of
the above form and to pull out c1 and c2 to pass to
fold_address_value_comparison. This allows us to reuse nearly all of the
existing folding logic while ensuring that we perform the comparison
correctly (on the address values) for capability pointer expressions.

gcc/ChangeLog:

	* fold-const.c (pointer_may_wrap_p): Adjust to handle capability
	pointers.
	(fold_address_value_comparison): New. Use it ...
	(fold_comparison): ... here.