contrib: use grep -E instead of egrep
egrep has been deprecated in favor of grep -E for a long time, and the
next grep release (3.8 or 4.0) will print a warning of egrep is used.
Stop using egrep so we won't see the warning.

Some non-POSIX grep implementations does not support -E option, so for
test_summary and warn_summary we prefer grep -E, but fallback to egrep
if grep -E does not work.  For check_GNU_style.sh, I think it is only
used by developers so hard coding grep -E is OK.

contrib/ChangeLog:

	* check_GNU_style.sh: Use grep -E instead of egrep.
	* test_summary: Use grep -E instead of egrep if it works.
	* warn_summary: Likewise.