fixincludes: use grep instead of egrep/fgrep
egrep/fgrep has been deprecated in favor of grep -E/-F for a long time,
and the next grep release (3.8 or 4.0) will print a warning if egrep or
fgrep is used.  Stop using egrep and fgrep so we won't see the warning.

But, we can't simply replace egrep to grep -E or fgrep to grep -F or the
build will break with some non-GNU grep implementations (lacking -E or
-F support).  autoconf documentation suggests to use AC_PROG_EGREP and
$EGREP or AC_PROG_FGREP and FGREP, but doing so is too complicated for
fixincludes.

So we simply adjust the patterns for a plain grep, instead of relying on
the behavior of grep -E/-F.

Q: Why "[ \t][ \t]*" instead of "[ \t]\\+"?
A: POSIX does not allow escaping + in BRE.

fixincludes/ChangeLog:

	* fixinc.in: Use grep instead of egrep.
	* inclhack.def: Use grep instead of egrep, and adjust an
	regular expression to make it a BRE.
	* genfixes: Use grep instead of egrep, and escape "." in the
	pattern.
	* fixincl.x: Regenerate.