Make -mblock-ops-vector-pair option valid as an ISA option.
I needed to turn off using load vector pair and store vector pair
instructions on the power10 for performance reasons.  It is likely that in
future machines, we will want to enable these instructions for block moves
and copies.  This patch adds the -mblock-ops-vector-pair option mask to
POWERPC_MASKS so that it is automatically set and cleared as we change the
target cpu.

2022-07-21   Michael Meissner  <meissner@linux.ibm.com>

gcc/

	* config/rs6000/rs6000-cpus.def (POWERPC_MASKS): Add
	OPTION_MASK_BLOCK_OPS_VECTOR_PAIR.
	* config/rs6000/rs6000.opt (-mblock-ops-vector-pair): Add a
	comment.