recog.c: Adjust find_constant_term_loc to handle capabilities
The code in recog.c:offsettable_address_addr_space_p used
find_constant_term_loc and then did:

*y2 = plus_constant (address_mode, *y2, mode_sz - 1);

on the result. The problem here is that *y2 may not be in address_mode.
Indeed, for capabilities, if we pull out a const_int in y2, then the
result will likely be in offset_mode (address_mode) instead.

This patch adjusts find_constant_term_loc, adding a new default-NULL
parameter which returns the mode of the result. In order to handle
modeless const_ints, we pass down the expected mode of the expression
from the context.

This fixes an ICE seen in g++.dg/torture/pr56302.C.

gcc/ChangeLog:

	* recog.c (find_constant_term_loc): Add mode_out parameter,
	implement by calling new find_constant_term_loc_1 function.
	(find_constant_term_loc_1): New.
	(offsettable_address_addr_space_p): Use correct mode when
	invoking plus_constant on result of find_constant_term_loc.
	* rtl.h (find_constant_term_loc): Add new default-NULL
	parameter through which the mode of the result can be
	returned.