[RISC-V] Fix 32bit riscv with zbs extension enabled
The problem here was a disconnect between splittable_const_int_operand
predicate and the function riscv_build_integer_1 for 32bits with zbs enabled.
The splittable_const_int_operand predicate had a check for TARGET_64BIT which
was not needed so this patch removed it.

Committed as obvious after a build for risc32-elf configured with --with-arch=rv32imac_zba_zbb_zbc_zbs.

Thanks,
Andrew Pinski

gcc/ChangeLog:

        * config/riscv/predicates.md (splittable_const_int_operand):
        Remove the check for TARGET_64BIT for single bit const values.

Signed-off-by: Andrew Pinski <apinski@marvell.com>
Change-Id: Id542ec511851993333a88fac54f78a04dcf2ca3a