Add predicates to inter-mix 128-bit floating point modes.
This patch adds two new predicates (ieee128_operand and ibm128_operand) that
allow KFmode and TFmode to be used inter-changeably when long double is IEEE
128-bit, and IFmode and TFmode to be used inter-changeabily with long double is
IBM 128-bit.

The various built-in functions that use KFmode or IFmode have been modified to
use these new predicates.

The code in rs6000_expand_builtin that switched between the KFmode and TFmode
built-in functions and the code that switched between the IFmode and TFmode
built-in functions has been modified so it looks at the argument type.  If a
TFmode argument was passed, it will convert the built-in to from using KFmode or
IFmode to one using TFmode.

The nop conversion functions between IFmode and TFmode when long double uses IBM
128-bit have been extended so that they take GPR registers as well as FPR
registers.  This is to allow the test "pr105334.c" to pass.  This test wants to
use -msoft-float along with the __ibm128 pack/unpack functions.

2022-08-11   Michael Meissner  <meissner@linux.ibm.com>

gcc/

	* config/rs6000/predicates.md (ieee128_operand): New predicate.
	(ibm128_operand): Likewise.
	* config/rs6000/rs6000-builtin.cc (rs6000_expand_builtin): Rework code
	that switches KFmode and IFmode built-in functions to TFmode to look at
	the argument's mode, and not just use the switch.
	* config/rs6000/rs6000.md (extendkftf2_internal): Add support for using
	GPR registers.
	(extendtfkf2_internal): Likewise.
	(extendiftf2_internal): Likewise.
	(extendtfif2_internal): Likewise.
	(unpack<mode>): Use ibm128_operand for TFmode/IFmode operands.
	(unpack<mode>_dm): Likewise.
	(unpack<mode>_nodm): Likewise.
	(pack<mode>_hard): Likewise.
	(unpack<mode>): Likewise.
	(pack<mode>): Likewise.
	(trunc<mode>sf2_hw): Use ieee128_operand for TFmode/KFmode operands.
	(add<mode>3_odd): Likewise.
	(sub<mode>3_odd): Likewise.
	(mul<mode>3_odd): Likewise.
	(div<mode>3_odd): Likewise.
	(sqrt<mode>2_odd): Likewise.
	(fma<mode>4_odd): Likewise.
	(fms<mode>4_odd): Likewise.
	(nfma<mode>4_odd): Likewise.
	(nfms<mode>4_odd): Likewise.
	(trunc<mode>df2_odd): Likewise.
	(cmp<mode>_hw): Likewise.
	* config/rs6000/vsx.md (xsxexpqp_<mode>): Likewise.
	(xsxsigqp_<mode>): Likewise.
	(xsiexpqpf_<mode): Likewise.
	(xsiexpqp_<mode>): Likewise.
	(xscmpexpqp_<code>_<mode>): Likewise.
	(xscmpexpqp_<code>_<mode>): Likewise.
	(xststdcqp_<mode>): Likewise.
	(xststdcnegqp_<mode): Likewise.
	(xststdcqp_<mode): Likewise.