gimple-fold: Fix folding of capability memcpy
Prior to this patch, the following C++ code was being miscompiled on
Morello:

void f(char *p, char *q)
{
  __builtin_memcpy (p, q, 16);
}

by being lowered to an __int128 load and store by
gimple_fold_builtin_memory_op. Inlining the memcpy like this is not
capability-preserving, but the copy is (just) big enough to potentially
be a capability copy.

In this patch, we adjust gimple_fold_builtin_memory_op to avoid inlining
__builtin_memcpy using integer loads/stores when we're on a capability
target and the copy is big enough to carry a capability on that target.