Add ieee128_operand.
Since we have 2 modes for IEEE 128-bit (KFmode and TFmode if long double is IEEE
128-bit), this patch provides a new predicate (ieee128_operand) that allows
combining a conversion from one mode to another.

For KFmode functions, we were are wanting a TFmode result (when long double uses
IEEE 128-bit), switch the built-in function to the same function using TFmode.

For IFmode functions, we were are wanting a TFmode result (when long double uses
IBM 128-bit), switch the built-in function to the same function using TFmode.

2022-08-12   Michael Meissner  <meissner@linux.ibm.com>

gcc/

	* config/rs6000/predicates.md (ieee128_operand): New predicate.
	* config/rs6000/rs6000.md (trunc<mode>sf2_hw): Use ieee128_operand.
	(round<mode>2): Likewise.
	(add<mode>3_odd): Likewise.
	(sub<mode>3_odd): Likewise.
	(mul<mode>3_odd): Likewise.
	(div<mode>3_odd): Likewise.
	(sqrt<mode>2_odd): Likewise.
	(fma<mode>4_odd): Likewise.
	(fms<mode>4_odd): Likewise.
	(nfma<mode>4_odd): Likewise.
	(nfms<mode>4_odd): Likewise.
	(trunc<mode>df2_odd): Likewise.
	* config/rs6000/vsx.md (xsxexpqp_<mode>): Likewise.
	(xsxsigqp_<mode>): Likewise.
	(xsiexpqpf_<mode>): Likewise.
	(xscmpexpqp_<code>_<mode): Likewise.
	(xscmpexpqp_<code>_<mode>): Likewise.
	(xststdcqp_<mode>): Likewise.
	(xststdcnegqp_<mode>): Likewise.
	(xststdcqp_<mode>): Likewise.