Update switching between KFmode to TFmode and IFmode to TFmode builtins
This patch looks to see if the caller function wants TFmode (when long double is
IEEE 128-bit) to switch the KFmode built-in functions to TFmode.

It will look to see if the caller functions wants IFmode (when long double is
IBM 128-bit) to switch the unpackif built-in function to unpacktf.

It will look to see if a TFmode argument is passed (when long double is IBM
128-bit) to switch the packif built-in function to packtf.

2022-08-12   Michael Meissner  <meissner@linux.ibm.com>

gcc/

	* config/rs6000/rs6000-builtin.cc (rs6000_expand_builtin): Update
	switching from KFmode/IFmode built-in functions to TFmode built-in
	functions.