Make __float128 use the _Float128 type.
Currently GCC uses the long double type node for __float128 if long double is
IEEE 128-bit.  It did not use the node for _Float128.

Problems showed up if you call the nansq function to make a signaling NaN (nansq
is mapped to nansf128).  Because the type node for _Float128 is different from
__float128, the machine independent code converts signaling NaNs to quiet NaNs
if the types are  not compatible.  The following tests used to fail when run on
a system where long double is IEEE 128-bit:

	gcc.dg/torture/float128-nan.c
	gcc.target/powerpc/nan128-1.c

This patch makes both __float128 and _Float128 use the same type node.

This patch also looks to see if the caller function was passed TFmode.  If it
was passed TFmode and the built-in function uses KFmode, switch the built-in
function to the equivalent TFmode built-in.  Similarly, if it was passed TFmode
and the built-in functions uses IFmode, switch the built-in function to the
equivalent TFmode built-in.

In the past the code would do the conversion based on what 128-bit floating
point format was used.  This patch changes this to look at the 	arguments.  If
any of the arguments use TFmode instead of KFmode or IFmode, it will do the
conversion.

Just in case the mode of the argument is different when we expand the built-in
function, do a convert_move instead of copy_to_mode_reg, which will abort
because the modes are different.

2022-08-17   Michael Meissner  <meissner@linux.ibm.com>

gcc/

	* config/rs6000/rs6000-builtin.cc (rs6000_init_builtins): Always use the
	_Float128 type for __float128.  Don't use the long double type.
	(rs6000_expand_builtin): Look at the argument types to see if a TFmode
	argument is passed to convert a KFmode or IFmode builtin to TFmode,
	instead of just looking at the options.
	* config/rs6000/rs6000.cc (rs6000_libgcc_floating_mode_supported_p):
	_Float128/__float128 types always use KFmode now.
	(rs6000_floatn_mode): Likewise.